tests: restructure tests to mirror respective src files and reformat code
