--- unittest/helper/common.py
@@ -1,3 +1,0 @@
-from pathlib import Path
-
-CONFIG_FILE = Path(__file__).parent.parent.resolve() / "testconfig_key.ini"
